ci: 首轮编译修复 + 头文件诊断输出
已确证修复:
- bsp_display: LCD_COLOR_FMT_RGB666 -> RGB888 (esp_lcd 无 RGB666)
- bsp_codec: audio_codec_i2c_cfg_t 去掉 .port (2.x 无此成员)
- set_out_vol 参数 float -> int (v1.6+/v2 均为 int)

诊断: CI command 中打印 managed_components 内 es8389_codec.h /
esp_codec_dev_*.h 内容（前 20KB），用于精确适配 2.0.0-beta3 的
es8389_codec_cfg_t 字段名。

+++ components/player/player.c
@@ -248,7 +248,7 @@ esp_err_t player_set_volume(uint8_t pct)
     if (s_codec == NULL) {
         return ESP_ERR_NOT_FOUND;
     }
-    if (esp_codec_dev_set_out_vol(s_codec, (float)pct) != ESP_CODEC_DEV_OK) {
+    if (esp_codec_dev_set_out_vol(s_codec, (int)pct) != ESP_CODEC_DEV_OK) {
         return ESP_FAIL;
     }
     return ESP_OK;
